+++ include/sys/timer.h
@@ -6,7 +6,7 @@
 #ifndef TIMER_H
 #define TIMER_H
 
-#include <stdint.h>
+#include <stddef.h>
 
 #define MAX_N_TICKS_TIMERS 5
 
@@ -16,7 +16,7 @@ struct n_ticks_timer {
 };
 
 void tick();
-uint32_t get_loops_per_tick();
+size_t get_loops_per_tick();
 int register_n_ticks_timer(int n_ticks, void (*timer_callback)());
 void busy_wait(int ticks_to_wait);
